+++ commands/errors.go
@@ -94,6 +94,7 @@ var (
 	ErrIsDisabled                                      = errors.New("is disabled")
 	ErrMustBeAtMost250                                 = errors.New("must be at most 250")
 	ErrNoUpdatesProvided                               = errors.New("no updates provided")
+	ErrMaxPriceMustRespectTickSize                     = errors.New("must respect tick size")
 )
 
 type Errors map[string][]error