+++ core/risk/margins_calculation.go
@@ -46,6 +46,19 @@ func scalingFactorsUintFromDecimals(sf *types.ScalingFactors) *scalingFactorsUin
 	}
 }
 
+func newMarginLevelsFull(maintenance num.Decimal) *types.MarginLevels {
+	maint, _ := num.UintFromDecimal(maintenance.Ceil())
+	return &types.MarginLevels{
+		MaintenanceMargin:      maint,
+		SearchLevel:            maint.Clone(),
+		InitialMargin:          maint.Clone(),
+		CollateralReleaseLevel: maint.Clone(),
+		OrderMargin:            num.UintZero(),
+		MarginMode:             types.MarginModeCrossMargin,
+		MarginFactor:           num.DecimalZero(),
+	}
+}
+
 func newMarginLevels(maintenance num.Decimal, scalingFactors *scalingFactorsUint) *types.MarginLevels {
 	umaintenance, _ := num.UintFromDecimal(maintenance.Ceil())
 	return &types.MarginLevels{
@@ -59,9 +72,85 @@ func newMarginLevels(maintenance num.Decimal, scalingFactors *scalingFactorsUint
 	}
 }
 
+func (e *Engine) calculateFullCollatMargins(m events.Margin, price *num.Uint, _ types.RiskFactor, withPotential bool) *types.MarginLevels {
+	var (
+		marginMaintenanceLng num.Decimal
+		marginMaintenanceSht num.Decimal
+	)
+	// convert volumn to a decimal number from a * 10^pdp
+	openVolume := num.DecimalFromInt64(m.Size()).Div(e.positionFactor)
+	aep := m.AverageEntryPrice().ToDecimal()
+	base := price.ToDecimal()
+	var (
+		riskiestLng = openVolume
+		riskiestSht = openVolume
+	)
+	// the party has no open positions that we need to calculate margin for
+	if riskiestLng.IsZero() && riskiestSht.IsZero() && !withPotential {
+		return &types.MarginLevels{
+			MaintenanceMargin:      num.UintZero(),
+			SearchLevel:            num.UintZero(),
+			InitialMargin:          num.UintZero(),
+			CollateralReleaseLevel: num.UintZero(),
+			OrderMargin:            num.UintZero(),
+			MarginMode:             types.MarginModeCrossMargin,
+			MarginFactor:           num.DecimalZero(),
+		}
+	}
+	if riskiestLng.IsPositive() {
+		marginMaintenanceLng = riskiestLng.Mul(aep)
+	} else {
+		// even our riskies position is short, get the sell AEP
+		marginMaintenanceLng = riskiestLng.Mul(base.Sub(aep)).Abs()
+	}
+	if riskiestSht.IsNegative() {
+		marginMaintenanceSht = riskiestSht.Mul(base.Sub(aep)).Abs()
+	} else {
+		// even the shortest position is long, get buy AEP
+		marginMaintenanceSht = riskiestSht.Mul(aep)
+	}
+	if withPotential {
+		// add margins required to cover the buy and sell orders
+		longSize := num.DecimalFromInt64(m.Buy()).Div(e.positionFactor)
+		// size * order price
+		longMargin := longSize.Mul(m.VWBuy().ToDecimal())
+		// add limit price * size to the margin required
+		shortSize := num.DecimalFromInt64(m.Sell()).Div(e.positionFactor).Abs()
+		// size * (max price - order price)
+		shortMargin := shortSize.Mul(base.Sub(m.VWSell().ToDecimal()))
+		marginMaintenanceLng = marginMaintenanceLng.Add(longMargin)
+		marginMaintenanceSht = marginMaintenanceSht.Add(shortMargin)
+	}
+	// now get the max margin required
+	if marginMaintenanceLng.GreaterThan(marginMaintenanceSht) {
+		return newMarginLevelsFull(marginMaintenanceLng)
+	}
+	// if short margin level > 0
+	if !marginMaintenanceSht.IsZero() {
+		return newMarginLevelsFull(marginMaintenanceSht)
+	}
+
+	// long margin level <= short, and short is zero, so no margin required.
+	return &types.MarginLevels{
+		MaintenanceMargin:      num.UintZero(),
+		SearchLevel:            num.UintZero(),
+		InitialMargin:          num.UintZero(),
+		CollateralReleaseLevel: num.UintZero(),
+		OrderMargin:            num.UintZero(),
+		MarginMode:             types.MarginModeCrossMargin,
+		MarginFactor:           num.DecimalZero(),
+	}
+}
+
 // Implementation of the margin calculator per specs:
 // https://github.com/vegaprotocol/product/blob/master/specs/0019-margin-calculator.md
 func (e *Engine) calculateMargins(m events.Margin, markPrice *num.Uint, rf types.RiskFactor, withPotentialBuyAndSell, auction bool, inc num.Decimal, auctionPrice *num.Uint) *types.MarginLevels {
+	if e.marginCalculator.FullyCollateralised {
+		if auction && auctionPrice != nil {
+			return e.calculateFullCollatMargins(m, auctionPrice, rf, withPotentialBuyAndSell)
+		}
+		return e.calculateFullCollatMargins(m, markPrice, rf, withPotentialBuyAndSell)
+	}
 	var (
 		marginMaintenanceLng num.Decimal
 		marginMaintenanceSht num.Decimal